requirements.txt
pytest
luigi
geojson
shapely
lxml
numpy
matplotlib
sklearn
Cython
tsne
annoy
marisa-trie
colour
falcon
falcon_multipart
cairocffi
jinja2
pygsp
pandas
gunicorn
palettable
python-daemon
requests